config.[ch]: provide config_size_t function and use it
For Windows compatibility.
cache.h: big_file_threshold & pack_size_limit_cfg are potentially size_t,
plus a few others convereted in this pass.

Other potential >4Gb variables are left for others.

+++ cache.h
@@ -865,8 +865,8 @@ extern int pack_compression_level;
 extern size_t packed_git_window_size;
 extern size_t packed_git_limit;
 extern size_t delta_base_cache_limit;
-extern unsigned long big_file_threshold;
-extern unsigned long pack_size_limit_cfg;
+extern size_t big_file_threshold;
+extern size_t pack_size_limit_cfg;
 
 /*
  * Accessors for the core.sharedrepository config which lazy-load the value
